pyproject.toml
[project]
name = "videoseo"
version = "0.1.0"
description = "An innovative platform designed to revolutionize the way video content is optimized for search engines and user engagement."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "colorlog>=6.10.1",
    "dotenv>=0.9.9",
    "fastapi>=0.121.2",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "strawberry-graphql[fastapi]>=0.285.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]